refactor: add get-wallet-address command helper
